requirements.txt
numpy>=1.24
scipy>=1.10
opencv-python>=4.8
scikit-image>=0.21
Pillow>=10.0
matplotlib>=3.7
PyYAML>=6.0
tqdm>=4.66
Flask>=3.0
Jinja2>=3.1

